Couple of small changes to get sim working in local environment
1. *Problem*: In admin_page.php, $gameInfo is initialized to NULL. We 
then try to access $gameInfo[“market_struct”] within the bodies of 
toggleSession and editGame. This causes an error if $gameInfo is NULL.  
*Solution*: when printing $gameInfo[“market_struct”] from php into the 
JS functions, included an ‘isset($gameInfo[“market_struct”]) && ’ 
immediately preceding. Now, if $gameInfo is NULL, we print FALSE into JS 
instead of a PHP error message (763 and 613)

2. *Problem*: In add_course.php, line 29 is missing a closing 
parenthesis  
*Solution*: added closing parthenthesis