Can remove games/courses and update games
The user can now remove games from the game options view and courses 
from the games view without error and are returned to the games or 
courses view, respectively.

The user can now update a game's settings from the game options view and 
are returned to the games view.

[*view this text in a markdown renderer for best reading 
experience*](https://dillinger.io/)


**Issues**
1. **Problem:** Presented with blank screen after deleting a game.
**Solution:** Changed the POST form field containing the course id of 
the game to delete submitted by admin_page.php from "deletedGameCourse" 
to "course_id" to allow uniform routing back to the games view from 
game_util.php (i.e. to admin_page.php with course query parameter).

2. **Problem:** add_course.php uses raw SQL to manually perform the 
deletion of a course.
**Solution:** Now, add_course.php makes a call to the 
$QW_DAO->deleteCourse method to handle the SQL and database 
interactions.

3. **Problem:** The QW_DAO->deleteCourse method performs a rowDie call 
to PDOX. The delete SQL command, however, does not return any rows upon 
a successful deletion, causing errors on every deletion attempt.
**Solution:** The QW_DAO->deleteCourse method now performs a queryDie to 
PDOX, which will crash if the query fails but not expect any row output 
if the query succeeds. This is the desired behavior in this case.

4. **Problem:** Similar problems to those detailed in the last commit 
for adding a game exist for updating a game. The QW_DAO->updateGame call 
didn't match what the method's definition expected.
**Solution:** First, I refactored QW_DAO->updateGame method call (in 
game_util.php) and definition (in QW_DAO.php) to be readable. I also 
resolved the ambiguity occuring between course and game IDs (as a 
stylistic choice, it might be best to rename the id column in the games 
and courses tables to game_id and course_id, respectively). Finally, I 
made sure the method call and definition agreed with each other and the 
games DB table.

**Adjustments**
1. The form submitted to delete a course (to add_course.php) and the 
form submitted to delete a game (to game_util.php) both use "deleteId" 
as the name of the field that contains the id of the course or game to 
be deleted, respectively. For clarity, the form submitted to delete a 
course or game now contains a field named "delete_course_id" or 
"delete_game_id" containing the id of the course or game to delete, 
respectively.

2. In add_course.php, we now double check that 'delete_course_id' is in 
the POST data before attempting to use it to delete a course

3. Changed the name of the field containing the game id in the hidden 
form submitted to update a game from "gameId" to "game_id".